Incorporate opponent play as an option to sb3_ppo agent
